Bio_precision_agent_V5/requirements.txt
streamlit>=1.30.0
openai>=1.0.0
duckduckgo-search>=6.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
markdown-pdf>=1.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
PyPDF2>=3.0.0
biopython>=1.83
nbformat>=5.9.2
